+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Changelog
 
+## [1.2.0] - 2026-06-17
+### Added
+- `cache.pages.maxAge` option to send `Cache-Control: max-age=<seconds>` for SiteGround Dynamic Cache.
+- Support for Kirby's native `cache.pages.ignore` option to send `Cache-Control: no-cache` for ignored pages.
+
+
 ## [1.1.0 - 1.1.2] - 2024-08-09
 ### Added
-- `/flush-cache` route to purge the cache (fo logged-in users only)
+- `/flush-cache` route to purge the cache (for logged-in users only)
 
 
 ## [1.0.0] - 2023-02-20